python/requirements.txt
python-dotenv==1.0.1
openai==1.3.0
requests==2.32.3
azure-search-documents==11.5.3
streamlit==1.32.0
